codex-tools: extract responses API tool models
